+++ src/languages/shell.ts
@@ -1,14 +1,15 @@
 import { ChildProcessWithoutNullStreams, spawn } from "child_process";
 import { writeFileSync, chmodSync, mkdirSync } from "fs";
 import { getTempPath } from "../config";
-import { Cell } from "../kernel";
 import * as vscode from "vscode";
 import path from "path";
 import { NotebookCell } from "vscode";
+import { CommentDecorator } from "../types";
+import { env_before, env_after } from "./shell-scripts"
 
 let tempDir = getTempPath();
 
-export const processShell = (cell: NotebookCell, language: string): ChildProcessWithoutNullStreams => {
+export const processShell = (cell: NotebookCell, language: string): {stream: ChildProcessWithoutNullStreams, clearOutput: boolean } => {
     let prog = ""
     switch(language){
         case "nushell":
@@ -30,19 +31,33 @@ export const processShell = (cell: NotebookCell, language: string): ChildProcess
     }
     let main = "";
     // Ignore all the clutter from the generated files when running tree
-    let contents = cell.document.getText().trim();
-    if (contents.endsWith("tree")) {
-        contents = "tree -I '__pycache__|main.sh|main.fish|main.nu|target'"
+    let contents = cell.document.getText();
+    
+    if (contents.trim() == "tree") {
+        writeFileSync(path.join(tempDir, ".gitignore"), `env_before.txt
+env_after.txt
+env_changes.sh
+main
+__pycache__
+rust/target
+venv
+        `);
+        contents = "tree --gitignore"
     }
-    main += `#!/bin/${language}\necho '!!output-start-cell'\n`;
-    main += contents;
 
+    // Save and load env vars on each shell incarnation
+    main += env_before + contents + env_after
+
+    let clearOutput = false;
+    if(contents.startsWith("# " + CommentDecorator.clear)){
+        clearOutput = true
+    }
 
     const filename = path.join(tempDir, `main`);
     mkdirSync(tempDir, { recursive: true });
     writeFileSync(filename, main);
     chmodSync(filename, 0o755);
 
-    return spawn(prog, [filename], {cwd: tempDir});
+    return {stream: spawn(prog, [filename], {cwd: tempDir}), clearOutput};
 };